feat(m2.9): extract Analytics domain bounded context with zero functional impact
