+++ core/src/services/location/LocationAnalyticsService.ts
@@ -1,112 +1,5 @@
-import {
-    LocationAnalytics,
-    logger,
-    AppError,
-    asString,
-    coerceLocationInput,
-    extractObjectIdString,
-    LOCATION_POPULARITY_WEIGHTS,
-    HOT_ZONE_SEARCH_THRESHOLD,
-    HOT_ZONE_ADS_THRESHOLD,
-    toLocationObjectId,
-    getActiveLocationById
-} from './_shared/locationServiceBase';
-import type { LocationAnalyticsEventType } from './_shared/locationServiceBase';
-export { normalizeGeoPoint, normalizeCoordinates } from './_shared/locationServiceBase';
-export const touchLocationAnalytics = async (
-    locationId: unknown,
-    eventType: LocationAnalyticsEventType,
-    increment = 1
-): Promise<void> => {
-    const objectId = toLocationObjectId(locationId);
-    if (!objectId || !Number.isFinite(increment) || increment <= 0) return;
-
-    const fieldMap: Record<LocationAnalyticsEventType, 'searchCount' | 'viewCount' | 'adsCount'> = {
-        location_search: 'searchCount',
-        ad_view: 'viewCount',
-        ad_post: 'adsCount'
-    };
-    const metricField = fieldMap[eventType];
-    const now = new Date();
-
-    await LocationAnalytics.collection.updateOne(
-        { locationId: objectId },
-        [
-            {
-                $set: {
-                    locationId: objectId,
-                    adsCount: { $ifNull: ['$adsCount', 0] },
-                    searchCount: { $ifNull: ['$searchCount', 0] },
-                    viewCount: { $ifNull: ['$viewCount', 0] }
-                }
-            },
-            {
-                $set: {
-                    [metricField]: { $add: [`$${metricField}`, increment] },
-                    lastUpdated: now
-                }
-            },
-            {
-                $set: {
-                    popularityScore: {
-                        $round: [
-                            {
-                                $add: [
-                                    { $multiply: [{ $ifNull: ['$adsCount', 0] }, LOCATION_POPULARITY_WEIGHTS.adsCount] },
-                                    { $multiply: [{ $ifNull: ['$searchCount', 0] }, LOCATION_POPULARITY_WEIGHTS.searchCount] },
-                                    { $multiply: [{ $ifNull: ['$viewCount', 0] }, LOCATION_POPULARITY_WEIGHTS.viewCount] }
-                                ]
-                            },
-                            2
-                        ]
-                    },
-                    isHotZone: {
-                        $or: [
-                            { $gte: [{ $ifNull: ['$searchCount', 0] }, HOT_ZONE_SEARCH_THRESHOLD] },
-                            { $gte: [{ $ifNull: ['$adsCount', 0] }, HOT_ZONE_ADS_THRESHOLD] }
-                        ]
-                    }
-                }
-            }
-        ],
-        { upsert: true }
-    );
-};
-
-export const touchLocationSearchAnalytics = async (
-    locationIds: Array<unknown>
-): Promise<void> => {
-    const unique = Array.from(
-        new Set(
-            locationIds
-                .map((value) => asString(value))
-                .filter((value): value is string => Boolean(value))
-        )
-    );
-    if (unique.length === 0) return;
-
-    await Promise.all(
-        unique.map((locationId) =>
-            touchLocationAnalytics(locationId, 'location_search', 1).catch((error) => {
-                logger.warn('Failed to track location search analytics', {
-                    locationId,
-                    error: error instanceof Error ? error.message : String(error)
-                });
-            })
-        )
-    );
-};
-
-export const logLocationEvent = async (payload: unknown) => {
-    const event = coerceLocationInput(payload);
-    const locationId = extractObjectIdString(event);
-    const type = asString((event as Record<string, unknown>).eventType) as LocationAnalyticsEventType | undefined;
-    if (locationId && type) {
-        const location = await getActiveLocationById(locationId);
-        if (!location) {
-            throw new AppError('Invalid or inactive location', 404, 'LOCATION_NOT_FOUND');
-        }
-        await touchLocationAnalytics(locationId, type, 1);
-    }
-    return true;
-};
+/**
+ * @deprecated Compatibility shim — import directly from domain:
+ * import { ... } from '@esparex/core/domains/analytics';
+ */
+export * from '../../domains/analytics/application/services/location/LocationAnalyticsService';
